# apps/dwa/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wellness Academy API",
    "version": "1.0.0",
    "description": "API for the Mental Health Education Platform — Wellness Academy. Session-based auth via Lucia, AI coaching via OpenRouter, content moderation on community features."
  },
  "servers": [
    {
      "url": "https://mental-health-education.soloframehub.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/ai/chat": {
      "post": {
        "summary": "Send a message to the wellness coach",
        "description": "Authenticated chat with the AI wellness coach. Includes crisis detection and personalized context from the user profile.",
        "tags": [
          "AI"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 4000
                  },
                  "history": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string",
                          "enum": [
                            "user",
                            "assistant"
                          ]
                        },
                        "content": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "role",
                        "content"
                      ]
                    },
                    "maxItems": 20,
                    "default": []
                  },
                  "context": {
                    "type": "object",
                    "properties": {
                      "courseId": {
                        "type": "string"
                      },
                      "lessonId": {
                        "type": "string"
                      },
                      "sectionId": {
                        "type": "string"
                      },
                      "pageContext": {
                        "type": "string"
                      }
                    }
                  }
                },
                "required": [
                  "message"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coach response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        },
                        "crisisDetected": {
                          "type": "boolean"
                        },
                        "crisisLevel": {
                          "type": "string",
                          "enum": [
                            "none",
                            "low",
                            "medium",
                            "high"
                          ]
                        }
                      },
                      "required": [
                        "message",
                        "crisisDetected"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limited — too many messages"
          },
          "503": {
            "description": "AI service unavailable"
          }
        }
      }
    },
...
  },
...
}